refactor: 서버에서 받는 데이터 모델 수정
