+++ OpenMarket/OpenMarket/ServerAccessRequest.swift
@@ -1,0 +1,49 @@
+//
+//  ServerAccessRequest.swift
+//  OpenMarket
+//
+//  Created by Sunny, James on 2021/05/11.
+//
+
+import Foundation
+
+struct MarketItemList: Decodable {
+    let page: Int
+    let items: [Item]
+}
+
+struct Item: Decodable {
+    let id: Int
+    let title: String
+    let price: Int
+    let currency: String
+    let stock: Int
+    let discountedPrice: Int?
+    let thumbnails: [String]
+    let registrationDate: Double
+    
+    private enum CodingKeys: String, CodingKey {
+        case id, title, price, currency, stock, thumbnails
+        case discountedPrice = "discounted_price"
+        case registrationDate = "registration_date"
+    }
+}
+
+struct ItemPage: Decodable {
+    let id: Int
+    let title: String
+    let descriptions: String
+    let price: Int
+    let currency: String
+    let stock: Int
+    let discountedPrice: Int?
+    let thumbnails: [String]
+    let images: [String]
+    let registrationDate: Double
+    
+    private enum CodingKeys: String, CodingKey {
+        case id, title, descriptions, price, currency, stock, thumbnails, images
+        case discountedPrice = "discounted_price"
+        case registrationDate = "registration_date"
+    }
+}
